bug#69171: Moving packages out of core to ELPA
> I agree.  I think this has even been suggested before, but I don't have the
> resources to take it on myself (though I'm happy to test it).  Maybe it's

Can you test the patch below (which basically applies the
diff between commits 627e0a14e87 to 77576cd7626e on `emacs.git).
Clearly, the code reorg made the work harder.

But it's a bit worse: the split chosen for the reorganization tends
to create circular dependencies (e.g. all the files need to depend on
`idlw-variables`, so `idlw-variables` can't depend on anything at all).
So it's harder to get rid of all the warnings.

I also attach the warnings resulting from compiling the code (after
applying the patch).

        Stefan